Migrate advxml.core.convert to advxml.core
