+++ src/main/scala/advxml/core/Converter.scala
@@ -1,8 +1,10 @@
-package advxml.core.convert
+package advxml.core
 
 import advxml.core.validate.ValidatedNelEx
 import cats.{Applicative, Id}
 import cats.data.Kleisli
+
+import scala.xml.NodeSeq
 
 /** Advxml
   * Created by geirolad on 31/10/2019.
@@ -46,7 +48,7 @@ object Converter {
   def apply[F[_], A, B](implicit F: Converter[F, A, B]): Converter[F, A, B] = F
 }
 
-private[convert] sealed abstract class FixedWrapperConverter[F[_]: Applicative, C[A, B] <: Converter[F, A, B]] {
+private[core] sealed abstract class FixedWrapperConverter[F[_]: Applicative, C[A, B] <: Converter[F, A, B]] {
 
   type Wrapper[_] = F[_]
 
@@ -60,4 +62,34 @@ private[convert] sealed abstract class FixedWrapperConverter[F[_]: Applicative, 
 }
 
 object PureConverter extends FixedWrapperConverter[Id, PureConverter]
+
 object ValidatedConverter extends FixedWrapperConverter[ValidatedNelEx, ValidatedConverter]
+
+object XmlConverter {
+
+  /** Syntactic sugar to convert a `O` instance into `X` using an implicit [[ModelToXml]] instance.
+    * This method catch a [[XmlToModel]] instance in the scope that conforms with types `X``O` and then invoke
+    * in it the method `apply` passing `xml`.
+    *
+    * @see [[ValidatedConverter]] for further information.
+    * @tparam X Contravariant input xml model type
+    * @tparam O Object output type
+    * @return [[advxml.core.validate.ValidatedNelEx]] instance that, if on success case contains `X` instance.
+    */
+  def asModel[X <: NodeSeq: XmlToModel[*, O], O](xml: X): ValidatedNelEx[O] = ValidatedConverter[X, O].run(xml)
+
+  /** Syntactic sugar to convert a `X` instance into `O` using an implicit [[XmlToModel]] instance.
+    *
+    * This method catch a [[ModelToXml]] instance in the scope that conforms with types `O``X` and then invoke
+    * in it the method `apply` passing `model`.
+    *
+    * See [[ModelToXml]] for further information.
+    *
+    * @see [[ValidatedConverter]] for further information.
+    * @tparam O Contravariant input model type
+    * @tparam X Output xml type
+    * @return [[advxml.core.validate.ValidatedNelEx]] instance that, if on success case contains `O` instance.
+    */
+  def asXml[O: ModelToXml[*, X], X <: NodeSeq](model: O): ValidatedNelEx[X] =
+    ValidatedConverter[O, X].run(model)
+}
