+++ src/test/scala/advxml/core/ValidatedConverterTest.scala
@@ -1,4 +1,4 @@
-package advxml.core.convert
+package advxml.core
 
 import advxml.core.validate.ValidatedNelEx
 import cats.data.Validated.Valid
